Initial implementation of memory contrained vtop
Currently performance is not impressive (60-100% of old vtop), because we are creating over 5 vectors that are heap allocated. Future optimizations will involve:
1. Using bumpalo to allocate a slice for all allocations to be contained within
2. Dropping bumpalo in 0.2, and just using a regular box

The wins of this implementation as it stands are: simplicity and memory usage reduction. While yes, work/wait stacks are incredibly fragile (one miscounted pop/push breaks the whole queue), the overall flow of translation is far more linear, with no need for search trees. Memory usage reduction is simply incredible. By using the stack system we are able to retrieve page map with just a few megabytes of allocation. Previously, we had to use a couple gigabytes, due to the approach being very much breadth-first.

All in all, this implementation is promising, and with heavy optimizations we should be able to exceed the performance of previous one.