Further optimizations for virtual translate
This particular commit implements a third item queue which is written to directly by split_chunk to avoid the need of popping addresses off the waiting stack. After a loop iteration, working_stack, and next_working_stack are swapped, and we continue onto the next loop. This increases memory requirements slightly (an additional working stack), but it is insignificant compared to the waiting stack. In addition, we could shrink the waiting stack by that amount and solve the regression.

A retest of old numbers is needed, due to Rust version change, but the current numbers are promising already. Peformance ranges from 25% slower to 30% faster (mostly on the slower side). Additional optimizations can, and will take place to bring performance up to speed.