More FFI reworkings. Make the plugin responsible for instance creation
Now the plugin provides a function create, which accepts arguments and spits out a PluginInstance upon success. This has the main advantage of object construction not being bound to specific type. This could allow to potentially return a optimized version of a kernel that performs address translations with ScopedVirtualTranslate of that specific architecture. Also, the proc_macro has become even shorter as a result.